Remove validation of IAM role from sharing_service.py and added checks in each s3 share operation
