+++ backend/dataall/modules/s3_datasets_shares/services/s3_share_service.py
@@ -4,6 +4,7 @@ from warnings import warn
 from dataall.base.db import utils
 from dataall.base.context import get_context
 from dataall.base.aws.sts import SessionHelper
+from dataall.base.aws.iam import IAM
 from dataall.core.permissions.services.resource_policy_service import ResourcePolicyService
 from dataall.core.permissions.services.tenant_policy_service import TenantPolicyService
 from dataall.core.environment.services.environment_service import EnvironmentService
@@ -11,6 +12,7 @@ from dataall.core.tasks.db.task_models import Task
 from dataall.core.tasks.service_handlers import Worker
 from dataall.modules.datasets_base.db.dataset_models import DatasetBase
 from dataall.modules.datasets_base.db.dataset_repositories import DatasetBaseRepository
+from dataall.modules.shares_base.db.share_object_models import ShareObject
 from dataall.modules.shares_base.db.share_object_repositories import ShareObjectRepository
 from dataall.modules.shares_base.db.share_state_machines_repositories import ShareStatusRepository
 from dataall.modules.shares_base.services.share_item_service import ShareItemService
@@ -272,3 +274,11 @@ class S3ShareService:
                 account_id=targetEnvAwsAccountId, database=old_shared_db_name, region=targetEnvRegion
             ).get_glue_database()
             return old_shared_db_name if database else datasetGlueDatabase + '_shared'
+
+    @staticmethod
+    def verify_principal_role(session, share: ShareObject) -> bool:
+        log.info('Verifying principal IAM role...')
+        role_name = share.principalRoleName
+        env = EnvironmentService.get_environment_by_uri(session, share.environmentUri)
+        principal_role = IAM.get_role_arn_by_name(account_id=env.AwsAccountId, region=env.region, role_name=role_name)
+        return principal_role is not None
