+++ backend/dataall/modules/s3_datasets_shares/services/share_processors/s3_bucket_share_processor.py
@@ -4,7 +4,7 @@ from typing import List
 
 from dataall.modules.shares_base.services.share_exceptions import PrincipalRoleNotFound
 from dataall.modules.s3_datasets_shares.services.share_managers import S3BucketShareManager
-from dataall.modules.shares_base.services.share_object_service import ShareObjectService
+from dataall.modules.s3_datasets_shares.services.s3_share_service import S3ShareService
 from dataall.modules.shares_base.services.shares_enums import (
     ShareItemHealthStatus,
     ShareItemStatus,
@@ -49,10 +49,10 @@ class ProcessS3BucketShare(SharesProcessorInterface):
         if not self.buckets:
             log.info('No Buckets to share. Skipping...')
             return success
-        if not ShareObjectService.verify_principal_role(self.session, self.share_data.share):
+        if not S3ShareService.verify_principal_role(self.session, self.share_data.share):
             raise PrincipalRoleNotFound(
                 'process approved shares',
-                f'Principal role {self.share_data.share.principalRoleName} is not found. Failed to update KMS key policy',
+                f'Principal role {self.share_data.share.principalRoleName} is not found. Failed to update KMS key/bucket policy',
             )
         for bucket in self.buckets:
             log.info(f'Sharing bucket {bucket.bucketUri}/{bucket.S3BucketName} ')
@@ -123,6 +123,11 @@ class ProcessS3BucketShare(SharesProcessorInterface):
             new_state = revoked_item_SM.run_transition(ShareObjectActions.Start.value)
             revoked_item_SM.update_state_single_item(self.session, removing_item, new_state)
             try:
+                if not S3ShareService.verify_principal_role(self.session, self.share_data.share):
+                    raise PrincipalRoleNotFound(
+                        'process revoked shares',
+                        f'Principal role {self.share_data.share.principalRoleName} is not found. Failed to update KMS key/bucket policy',
+                    )
                 manager.delete_target_role_bucket_policy()
                 manager.delete_target_role_access_policy(
                     share=self.share_data.share,
@@ -163,6 +168,11 @@ class ProcessS3BucketShare(SharesProcessorInterface):
                 bucket.bucketUri,
             )
             try:
+                if not S3ShareService.verify_principal_role(self.session, self.share_data.share):
+                    raise PrincipalRoleNotFound(
+                        'process verify shares',
+                        f'Share principal Role {self.share_data.share.principalRoleName} not found. Check the team or consumption IAM role used.',
+                    )
                 manager.check_role_bucket_policy()
                 manager.check_s3_iam_access()
